Fix ESPrima make rule
