+++ capture/shim.js
@@ -1,3 +1,3 @@
 function define(name, deps, fn) {
-    fn.apply(window, deps.map(function() { return window; }));
+    fn.apply(this, deps.map(function() { return this; }));
 }
